requirements.txt
boto3
pyyaml
tqdm
